+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Added
+- Add --assets-token and --assets-token-name arguments for fetching assets from private repositories
+
 ### Fixed
+- Fix asset names retrieval in URLs when uploading assets to Zenodo or RADAR
 - In `utils/metadata/codemeta.py` : fix bug in `remove_doubles` when author or creator fields contain a single entry of type dict.
 - In `utils/metadata/radar.py` : support case when author/contributor affiliation(s) is not a list.